Begin front end branch with App.js and NavBar.js components
We will attempt to split our main branches as front-end and back-end - we may ultimately need to merge them in order to test functionality as the API calls in the front end React will not resolve until schema, Seeder, and Routers are established. Locally-hosted elements such as photos will be added during HTML/CSS review so will show as broken links until added.